--wip-- work on nifti2, introduce base class for extension
